bidnamic/requirements.txt
asgiref==3.5.0
coverage==6.3.2
Django==3.2.12
django-crispy-forms==1.14.0
Faker==13.3.1
factory-boy==3.2.1
pre-commit==2.17.0
pytz==2021.3
python-dotenv==0.18.0
sqlparse==0.4.2
gunicorn==20.1.0
psycopg2_binary==2.9.2
